Fix: UserId 타입불일치 수정
.

+++ server/photomory/src/main/java/com/example/photomory/controller/OurAlbumController.java
@@ -40,7 +40,7 @@ public class OurAlbumController {
         }
 
         UserEntity user = (UserEntity) authentication.getPrincipal();
-        Integer userId = user.getUserId();
+        Long userId = user.getUserId();
 
         try {
             OurAlbumResponseDto response = ourAlbumService.getAlbumDetails(albumId, userId);
